fix: ensure all roles in roles collection
