+++ imports/plugins/core/accounts/server/no-meteor/util/defaultRoles.js
@@ -1,0 +1,44 @@
+export const defaultCustomerRoles = [
+  "account/profile",
+  "cart/completed",
+  "guest",
+  "index",
+  "product",
+  "tag"
+];
+
+export const defaultVisitorRoles = [
+  "anonymous",
+  "cart/completed",
+  "guest",
+  "index",
+  "product",
+  "tag"
+];
+
+export const defaultOwnerRoles = [
+  "account/profile",
+  "admin",
+  "cart/checkout",
+  "cart/completed",
+  "createProduct",
+  "dashboard",
+  "guest",
+  "index",
+  "owner",
+  "product",
+  "shopSettings",
+  "tag"
+];
+
+export const defaultShopManagerRoles = [
+  "account/profile",
+  "cart/completed",
+  "createProduct",
+  "dashboard",
+  "guest",
+  "index",
+  "product",
+  "shopSettings",
+  "tag"
+];
